fix: 🐛 page refreshing when not modified
This was done by moving to the Last-Modified header in favour of custom
behaviour, so the browser now does this automatically for us. In
addition, migrated to @vercel/speed-insights, and changed the image size
of TMDB poster images to 93 instead of 500 to reduce the amount of data
the user has to load, only for us to then shrink it wayyyy down.